Implementação dos métodos restantes da api.

+++ Repository/Order/RepOrder.cs
@@ -5,12 +5,14 @@ namespace repository
 {
     public class RepOrder : IRepOrder
     {
+        #region Ctor
         private readonly IMongoCollection<Order> _mongoCollection;
 
         public RepOrder(IMongoCollection<Order> mongoCollection)
         {
             _mongoCollection = mongoCollection;
         }
+        #endregion
 
         #region InsertOrder
         public Order InsertOrder(Order order)
@@ -34,6 +36,11 @@ namespace repository
         {
             var order = _mongoCollection.Find(p => p.Id == Id).FirstOrDefault();
 
+            if(order == null)
+            {
+                throw new Exception("Pedido não encontrado.");
+            }
+
             return order;
         }
         #endregion
@@ -50,7 +57,67 @@ namespace repository
         #region DeleteOrder
         public void DeleteOrder(int id)
         {
+            var order = _mongoCollection.Find(p => p.Id == id).FirstOrDefault();
+
+            if(order == null)
+            {
+                throw new Exception("Pedido informado não encontrado.");
+            }
+
             _mongoCollection.DeleteOne(p => p.Id == id);
+        }
+        #endregion
+
+        #region GetOrderByCustomerId
+        public List<Order> GetOrderByCustomerId(int customerId)
+        {
+            var orders = _mongoCollection.Find(p => p.CustomerId == customerId).ToList();
+
+            if(orders == null || orders.Count == 0)
+            {
+                throw new Exception("Cliente não possui pedidos.");
+            }
+
+            return orders;
+        }
+        #endregion
+
+        #region GetPriceTotalOrder
+        public PriceTotalOrderDTO GetPriceTotalOrder(int Id)
+        {
+            var order = _mongoCollection.Find(p => p.Id == Id).FirstOrDefault();
+
+            if(order == null)
+            {
+                throw new Exception("Pedido não encontrado.");
+            }
+
+            decimal total = order.Itens.Sum(item => item.Price);
+
+            return new PriceTotalOrderDTO
+            {
+                Id = order.Id,
+                Price = total
+            };
+        }
+        #endregion
+
+        #region OrderByCustomer
+        public OrderByCustomerDTO OrderByCustomer(int customerId)
+        {
+            var orders = _mongoCollection.Find(p => p.CustomerId == customerId).ToList();
+
+            if (orders == null || orders.Count == 0)
+            {
+                throw new Exception("Cliente não possui pedidos.");
+            }
+
+            return new OrderByCustomerDTO
+            {
+                Id = orders.Select(order => order.Id).ToList(),
+                QtdeOrders = orders.Count()
+            };
+
         }
         #endregion
     }
